==> DEL_Headspace/2.Headspace_Gas.R <==
# Analyze and plot headspace CO2 and CH4
# University of California, Berkeley
# Maggie Mengting Yuan
# CR1000 input file is generated by DEL_CO2.R under the same GitHub repository.
# Methane data is a subset from Picarro output, cleaned up following #insert Jordan Hoff GitHub page#).
# This code generates Figure 2 in the manuscript.

rm(list=ls())

library(tidyverse)
library(agricolae)
library(ggplot2)
library(gridExtra)

setwd("/Users/maggieyuan/Documents/GitHub/DynamicEcosystemLabeling_CO2_analysis/DEL_Headspace")

L3_data = read.table("Data/CR1000_L3_clean.dat", header=T)
CH4_data = read.table("Data/Picarro_CH4_Tenth.csv", sep=",", header=T) 

## match CH4 data to L3_data
L3_data$HR_12CH4 = unlist(
  lapply(L3_data$EpochTime, 
         FUN=function(x){
           id = which(abs(x-CH4_data$EPOCH_TIME)==min(abs(x-CH4_data$EPOCH_TIME)))[1]
           return(CH4_data$HR_12CH4[id])
           }
         )
  )

L3_data$HR_Delta_iCH4_Raw = unlist(
  lapply(L3_data$EpochTime, 
         FUN=function(x){
           id = which(abs(x-CH4_data$EPOCH_TIME)==min(abs(x-CH4_data$EPOCH_TIME)))[1]
           return(CH4_data$HR_Delta_iCH4_Raw[id])
           }
         )
  )

L3_data$Total_CH4 = unlist(
  lapply(L3_data$EpochTime, 
         FUN=function(x){
           id = which(abs(x-CH4_data$EPOCH_TIME)==min(abs(x-CH4_data$EPOCH_TIME)))[1]
           return(CH4_data$Total_CH4[id])
           }
         )
  )

L3_data %>% as.tibble

## plot head space gas
L3_axis_epoch = seq(from=1553007600, by=3600*24, length.out=13)
L3_axis_text = c("3/19/19\n-8AM", "3/20", "3/21", "3/22", "3/23", "3/24", "3/25", "3/26", "3/27", "3/28", "3/29", "3/30", "3/31")

# calculate stats across chambers
co2_aggregate_by_cycle = function(co2_data){
  time_cycle_ids = list() # need to group each cycle of P01 to P16 measurements in order to calculate mean and sd
  time_stamp = c()
  
  time_cycle_ids[[1]]=1
  time_stamp[[1]]=co2_data$TIMESTAMP[1]
  for (i in 2:length(co2_data$Plot)){
    if (co2_data$Plot[i]=="P01"){
      time_cycle_ids = c(time_cycle_ids, as.list(i))
      time_stamp = c(time_stamp, as.character(co2_data$TIMESTAMP[i])) # use time stamp of P01 for the entire cycle. Each cycle is 80min.
    }else{
      time_cycle_ids[[length(time_cycle_ids)]] = c(time_cycle_ids[[length(time_cycle_ids)]], i)
    }
  }
  
  time_cycle_mean = t(as.data.frame(lapply(time_cycle_ids, 
                                           FUN=function(row_ids, data){
                                             subdata = co2_data[row_ids,c("EpochTime", "Irga_CO2", "Picarro_CO2", "Picarro_Atm_Pct", "ws_Temp_C", "PAR_In", "HR_12CH4", "HR_Delta_iCH4_Raw", "Total_CH4")]
                                             return(colMeans(subdata, na.rm=T))}, 
                                           data=co2_data)))
  
  time_cycle_sd = t(as.data.frame(lapply(time_cycle_ids, 
                                         FUN=function(row_ids, data){
                                           subdata = co2_data[row_ids,c("Irga_CO2", "Picarro_CO2", "Picarro_Atm_Pct", "ws_Temp_C", "HR_12CH4", "HR_Delta_iCH4_Raw", "Total_CH4")]
                                           col_sd = apply(subdata, 2, FUN=sd, na.rm=T)
                                           return(col_sd)}, 
                                         data=co2_data)))
  
  time_cycle_se = t(as.data.frame(lapply(time_cycle_ids, 
                                         FUN=function(row_ids, data){
                                           subdata = co2_data[row_ids,c("Irga_CO2", "Picarro_CO2", "Picarro_Atm_Pct", "ws_Temp_C", "HR_12CH4", "HR_Delta_iCH4_Raw", "Total_CH4")]
                                           col_se = apply(subdata, 2, FUN=function(x){return(sd(x, na.rm=T)/sqrt(sum(!is.na(x))))})
                                           return(col_se)}, 
                                         data=co2_data)))
  
  row.names(time_cycle_mean) = time_stamp
  row.names(time_cycle_sd) = time_stamp
  row.names(time_cycle_se) = time_stamp
  
  output = list(time_cycle_mean, time_cycle_sd, time_cycle_se)
  names(output) = c("mean", "sd", "se")
  
  return(output)
}
L3_cycle = co2_aggregate_by_cycle(co2_data=L3_data)

# Figures across chambers, Figure 2 in manuscript

df.CO2_13C_chambers = data.frame(
  Time = L3_cycle[["mean"]][,"EpochTime"],
  CO2 = L3_cycle[["mean"]][,"Picarro_CO2"],
  lb = L3_cycle[["mean"]][,"Picarro_CO2"] - L3_cycle[["sd"]][,"Picarro_CO2"],
  ub = L3_cycle[["mean"]][,"Picarro_CO2"] + L3_cycle[["sd"]][,"Picarro_CO2"])

df.atm = data.frame(
  Time = L3_cycle[["mean"]][,"EpochTime"],
  atm = L3_cycle[["mean"]][,"Picarro_Atm_Pct"],
  lb = L3_cycle[["mean"]][,"Picarro_Atm_Pct"] - L3_cycle[["sd"]][,"Picarro_Atm_Pct"],
  ub = L3_cycle[["mean"]][,"Picarro_Atm_Pct"] + L3_cycle[["sd"]][,"Picarro_Atm_Pct"])

df.CH4 = data.frame(
  Time = L3_cycle[["mean"]][,"EpochTime"],
  CH4 = L3_cycle[["mean"]][,"Total_CH4"],
  lb = L3_cycle[["mean"]][,"Total_CH4"] - L3_cycle[["sd"]][,"Total_CH4"],
  ub = L3_cycle[["mean"]][,"Total_CH4"] + L3_cycle[["sd"]][,"Total_CH4"])

df.PAR = data.frame(
  Time = L3_cycle[["mean"]][,"EpochTime"],
  PAR = L3_cycle[["mean"]][,"PAR_In"])

p.CO2 = ggplot(df.CO2_13C_chambers, aes(Time)) + 
  geom_line(aes(y=CO2), colour="black") + 
  geom_point(aes(y=CO2), colour="black", size=0.8) + 
  geom_ribbon(aes(ymin=lb, ymax=ub), fill="black", alpha=0.3) +
  scale_x_continuous(breaks=L3_axis_epoch, labels=L3_axis_text) +
  ylab(expression(CO[2]~(ppm))) +
  ylim(-100,3000) +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank())

p.atm = ggplot(df.atm, aes(Time)) + 
  geom_line(aes(y=atm), colour="black") + 
  geom_point(aes(y=atm), colour="black", size=0.8) + 
  geom_ribbon(aes(ymin=lb, ymax=ub), fill="black", alpha=0.3) +
  scale_x_continuous(breaks=L3_axis_epoch, labels=L3_axis_text) +
  ylab(expression(paste("Atom%", " "^{13}, "C"))) +
  ylim(-10, 100) +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank())

p.CH4 = ggplot(df.CH4, aes(Time)) + 
  geom_line(aes(y=CH4), colour="black") + 
  geom_point(aes(y=CH4), colour="black", size=0.8) + 
  geom_ribbon(aes(ymin=lb, ymax=ub), fill="black", alpha=0.3) +
  scale_x_continuous(breaks=L3_axis_epoch, labels=L3_axis_text) +
  ylab("CH4 (ppm)") + 
  ylim(1.93,2.16) +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank())

p.PAR = ggplot(df.PAR, aes(Time)) + 
  geom_line(aes(y=PAR), colour="black") + 
  scale_x_continuous(breaks=L3_axis_epoch, labels=L3_axis_text) +
  ylab("PAR") + 
  ylim(0,2200) +
  xlab("Date")

grid.arrange(p1, p2, p3, p4, nrow=2)

#pdf("headspace_CO2_CH4_L3.pdf", width=6, height=6)
grid.arrange(
  grobs = list(p.CO2, p.atm, p.CH4, p.PAR),
  layout_matrix = rbind(1,
                        2,
                        3,
                        4)
)
#dev.off()


# stats
#Averaged across chambers, mean +/- se
L = as.data.frame(L3_cycle[["mean"]])

max(L[,"Picarro_Atm_Pct"][which(L[,"PAR_In"]>200)]) #Max daytime (PAR>200) 13C atom%
mean(L[,"Picarro_Atm_Pct"][which(L[,"PAR_In"]>200)]) #Average daytime (PAR>200) 13C atom%
sd(L[,"Picarro_Atm_Pct"][which(L[,"PAR_In"]>200)]) #sd=

mean(L[,"Picarro_CO2"][which(L[,"PAR_In"]>200)]) #Average daytime (PAR>200) CO2 ppm
sd(L[,"Picarro_CO2"][which(L[,"PAR_In"]>200)]) #sd

mean(L[,"Picarro_CO2"][which(L[,"PAR_In"]==0)]) #Average nightime (PAR=0) CO2 ppm
sd(L[,"Picarro_CO2"][which(L[,"PAR_In"]==0)]) #sd

#nightime (PAR=0) atom% 13C ppm
L_night = L[,c("EpochTime", "Picarro_Atm_Pct")][which(L[,"PAR_In"]==0),]
cycle = rep(NA, nrow(L_night))
cycle[1] <-1
for (i in 2:length(cycle)){
  if (L_night[i,"EpochTime"]-L_night[i-1,"EpochTime"] > 3600*3){
    cycle[i] = cycle[i-1]+1
  }else{
    cycle[i] = cycle[i-1]
  }
}
L_night = cbind(L_night, cycle)
L_nightly_13atm = aggregate(L_night[,"Picarro_Atm_Pct"], by=list(L_night[,"cycle"]), FUN=mean)
summary(lm(L_nightly_13atm$x ~ L_nightly_13atm$Group.1))

#Max, min, average, sd of air T
max(L[,"ws_Temp_C"])
min(L[,"ws_Temp_C"])
mean(L[,"ws_Temp_C"])
sd(L[,"ws_Temp_C"])
