==> DEL_Headspace/3.Reco.R <==
# Calculate overnight ecosystem respiration from headspace CO2 concentration
# University of California, Berkeley
# Maggie Mengting Yuan
# CR1000 input file is generated by DEL_CO2.R under the same GitHub repository.

# The chambers were kept closed after sunset, when CO2 supply stopped.
# The chambers were vented in the morning before labeling started.
# Overngith CO2 accumulated was used to estimate ecosystem respiration.
# Only 13C chambers were used, as CO2 accumulation reached maximum bound for IRGA (2000ppm) quickly for most plots. Picarro maxinum bound = 8000ppm.
# This involved manual removal of outliers (measurements deviated from an increasing curve).

# This code generates Figures 

rm(list=ls())
#library(agricolae)
#library(ggplot2)
#library(gridExtra)

## functions used

night_co2_by_plot = function(plot, labeling_data, axis_epoch=L3_axis_epoch, axis_text=L3_axis_text){ # plot:number of plot
  plot_name = paste("P", sprintf("%02d",plot), sep="")
  plot_data = labeling_data[which(labeling_data$Plot == plot_name),]
  
  plot_data $Picarro_CO2[is.na(plot_data $Picarro_CO2)] <-0
  plot_data $Picarro_Atm_Pct[is.na(plot_data $Picarro_Atm_Pct)] <-0
  
  plot_night = plot_data[which(plot_data$PAR_In==0),]	
  
  par(mfrow=c(1,2), oma=c(0.5,0.5,2,1), mar=c(2,5,1,0))
  plot(plot_night $EpochTime, plot_night $Picarro_CO2, cex=0.5, ylab="CO2 (ppm) in 13C chamber", xaxt="n")
  axis(1, at= axis_epoch, labels= axis_text)
  plot(plot_night $EpochTime, plot_night $Picarro_Atm_Pct, cex=0.5, ylab="Atom% 13C in 13C chamber", xaxt="n")
  axis(1, at= axis_epoch, labels= axis_text)
  mtext(plot_name, outer=T, cex=1.5)
}	# generate figures of nighttime headspace CO2 over time for each plot

output_night_CO2 = function(plot, labeling_data, labeling){
  plot_name = paste("P", sprintf("%02d",plot), sep="")
  plot_data = labeling_data[which(labeling_data$Plot == plot_name),]
  
  plot_night = plot_data[which(plot_data$PAR_In==0),]	
  plot_night$night_index = format(as.POSIXlt(plot_night$TIMESTAMP, format = "%Y-%m-%d %H:%M:%S") + 7*60*60, format = "%d")
  
  write.table(plot_night, paste("nightCO2/nightCO2_", labeling, "_", plot_name, ".txt", sep=""), sep="\t", row.names=F, quote=F)
} # generate tables for nighttime headsapce CO2 over time

night_co2_by_plot_curated = function(curated_co2_data, axis_epoch = L3_axis_epoch, axis_text = L3_axis_text){
  plot_name = paste("P", sprintf("%02d",i), sep="")
  plot_night = curated_co2_data[which(curated_co2_data$Picarro_CO2_Eyeball_MY=="Y"),]	
  plot(plot_night$EpochTime, plot_night$Picarro_CO2, cex=0.5, ylab="CO2 (ppm) in 13C chamber", xaxt="n", main=plot_name)
  axis(1, at=axis_epoch, labels=axis_text)
}	# generate figures of nighttime headspace CO2 over time for each plot, after removing outliers.

CO2_rate = function(curated_co2_data, chamber_vol=v_chamber, plot_area=a_plot){
  atm.pr = 101325 # atmospheric pressure in Pa
  gas.cst = 8.314 # gas constent in m3 * Pa * K-1 * mol-1
  
  curated_co2_data = curated_co2_data %>% mutate(ws_Temp_K = ws_Temp_C+273.15) %>% # calculate temperature in Kelvin
    mutate(n.co2 = atm.pr *chamber_vol /gas.cst /ws_Temp_K *Picarro_CO2/1000000) %>% # calculate amount of CO2 molecules
    mutate(n.12co2 = n.co2*(1-Picarro_Atm_Pct/100), n.13co2 = n.co2*Picarro_Atm_Pct/100) %>% # separate 12CO2 and 13CO2 molecules 
    mutate(m.c.ug = n.12co2*12*1000000 + n.13co2*13*1000000) # calculate mass of C in micrograms
  
  rate = lm(m.c.ug~EpochTime, data=curated_co2_data)$coefficients["EpochTime"]/plot_area # rate of respiration in ugC * m-2 * s-1
  return(rate)
}

#####################################################################
setwd("/Users/maggieyuan/Documents/GitHub/DynamicEcosystemLabeling_CO2_analysis/DEL_Headspace")

L3_data = read.table("Data/CR1000_L3_clean.dat", header=T)

v_chamber = pi*0.203^2*1.3  # calculate the volume of the chamber in m^3, for L1 to L3
a_plot = pi*0.203^2  # calculate the area of plot surface in m^2

L3_axis_epoch = seq(from=1553007600, by=3600*24, length.out=13)
L3_axis_text = c("3/19/19\n-8AM", "3/20", "3/21", "3/22", "3/23", "3/24", "3/25", "3/26", "3/27", "3/28", "3/29", "3/30", "3/31")

## 1. manual curation to remove outliers

for (i in 1:16){
  png(paste("nightCO2/L3_", i, ".png", sep=""), width=1500, height=300)
  night_co2_by_plot(plot=i, labeling_data=L3_data)
  dev.off()
}

for (i in 1:16){output_night_CO2(plot=i, labeling_data=L3_data, labeling="L3")}

# Then look and remove data points based on Figures generated. 
# Add a column in each data table called "Picarro_CO2_Eyeball_MY" to indicate if the datapoint shall be included (Y) or removed (N). Save as a curated file.

# plot out curated data and check again
for (i in 1:16){
  png(paste("nightCO2/L3_", i, ".curated.png", sep=""), width=700, height=300)
  night_co2_by_plot_curated(curated_co2_data = read.table(paste("nightCO2/nightCO2_L3_P", sprintf("%02d",i), ".curated.txt", sep=""), header=T, sep="\t"))
  dev.off()
}

## 2. calcualte nighttime Reco from curated data

# merge plot data into a single file
file_list <- list.files(path="nightCO2", pattern="curated.txt", full.names=T)
rm(list="night_co2")
for (file in file_list){
  # if the merged dataset doesn't exist, create it
  if (!exists("night_co2")){
    night_co2 <- read.table(file, header=TRUE, sep="\t")
  }
  
  # if the merged dataset does exist, append to it
  if (exists("night_co2")){
    temp_dataset <-read.table(file, header=TRUE, sep="\t")
    night_co2<-rbind(night_co2, temp_dataset)
    rm(temp_dataset)
  }
}
dim(night_co2)

# split data into sections on which CO2 change rate will be calculated
night_co2_group = night_co2 %>% 
  filter(Picarro_CO2_Eyeball_MY=="Y") %>% # remove outliers
  group_by(Plot, night_index) # group by plot and night index. In each element, CO2 flux rate will be calculated.

night_co2_rate = group_keys(night_co2_group) %>% mutate(rate = unlist(lapply(group_split(night_co2_group), FUN=CO2_rate))) # Night Reco in ugC.m-2.s-1

# Add plot treatment
info = read.table("Data/plot_trt.txt", header=T, sep="\t")
night_co2_rate = left_join(night_co2_rate, info)

## 3. stats
night_co2_aov = aov(rate ~ Water*as.factor(night_index), data=night_co2_rate)
summary(night_co2_aov)
night_co2_lsd = LSD.test(night_co2_aov, "Water", p.adj="holm")

## 4.output data
night_co2_rate_output = night_co2_rate %>% rename(Reco.ugC.s.m2 = rate) %>% mutate(Date = paste("3/",night_index,"/19",sep="")) %>% select(!night_index)
write.table(night_co2_rate_output, "Data/Night_Reco.txt", row.names=F, quote=F, sep="\t") # look at what columns to change
